refactor: move context menu to requests
